requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
langchain
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
sqlite3-dict
redis
python-multipart
aiofiles
httpx
asyncio-throttle
slowapi
prometheus-client
alembic
passlib[bcrypt]
python-jose[cryptography]
scikit-learn
plotly
seaborn
matplotlib
nltk
textblob
wordcloud
apscheduler
celery[redis]
pydantic[email]
email-validator 